Refactored and modified substantially the gui of the clients
There's now more shared code, and many things are implementation indipendent.
Toolbars live now in specific pannels, and there's now in the gui a side bar used (in future) for status notifications.
Last but not the least, size of the gui is now really determined by contents and centered everythime contents change.