ArticleGroup: added groups of articles support
this is much more serverside: in the cash-desk client is not used yet

Orders do not take advantage of it, too; since in the server there is not enough infrastructure to redirect groups of article in an order to specific clients that listen to such groups.

Added also table support in the DB and some internationalization fixes in the code.